requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cloudinary==1.26.0
Django==3.2.7
djangorestframework==3.12.4
et-xmlfile==1.1.0
importlib-metadata==4.8.1
Markdown==3.3.4
openpyxl==3.0.9
pkg-resources==0.0.0
psycopg2==2.9.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
urllib3==1.26.7
zipp==3.5.0
